requirements.txt
en-core-sci-lg==0.5.4
faiss-cpu==1.13.1
fhir-resources==8.1.0
litellm==1.66.0
mysql-connector-python==9.5.0
numpy==1.26.4
pydantic==2.11.1
python-dotenv==1.1.0
pyucum==0.1.3
scikit-learn==1.6.1
spacy==3.7.5
torch==2.9.1+cu126
tqdm==4.67.1
transformers==4.51.3
